--- src/controller/criarContaController.js
@@ -1,8 +1,0 @@
-const criarcontaController = {
-	criarConta: function(req, res, next) {
-		res.render('criarConta', { title: "Criar Conta" });
-	}
-}
-
-
-module.exports = criarcontaController
