--- src/controller/loginController.js
@@ -1,8 +1,0 @@
-const loginController = {
-    login: (req, res) => {
-        return res.render('login');
-    }
-};
-
-module.exports = loginController;
-
