+++ src/routes/clientes.js
@@ -3,5 +3,7 @@ const router = express.Router();
 const clientesController = require('../controller/clientesController');
 
 router.get('/', clientesController.list);
+router.get('/criarconta', clientesController.create);
+router.post('/', clientesController.store);
 
 module.exports = router;
