--- src/routes/login.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-
-const loginController = require('../controller/loginController');
-
-router.get("/", loginController.login);
-
-module.exports = router;
