+++ src/views/acessar.ejs
@@ -29,7 +29,7 @@
 
         <div class="entrar"><button type="submit">Entrar</button></div>
 
-        <div class="aqui">Não tem conta? <a href="/criarconta">Crie uma agora!</a></div>
+        <div class="aqui">Não tem conta? <a href="/clientes/criarconta">Crie uma agora!</a></div>
       </form>
     </div>
